Fully tested all behaviors
Wrote tests for each discrete behavior of the installer class. This
identified several regexp issues, as well as some bad conditionals, all
of which were fixed.

In order to test, I created a "functions" file in the same namespace as
the installer class, which redefines the `is_file()`,
`file_get_contents()` and `file_put_contents()` functions such that they
communicate via a static class; the tests then run assertions on the
contents of that class.