Implemented selection caching
One TODO item was to allow prompting to use the given selection for any
other packages installed in the same session.

Interestingly, Composer must run scripts in separate processes, because
the prompt value could not be cached between invocations in a static
property.

As such, this adds a new class, `InjectorCache`, which caches to a local
file the first time a selection is chosen to persist, and then reads
from that file on subsequent package installs. When the install session
is complete, it removes the file.