Refactor to collections to implement arrays of modules
In order to allow the `extra.zf.<type>` values to be either module names *or*
arrays of modules names, a number of changes were necessary. To reduce the
number of nested loops, this patch introduces a Collection type, used within the
ComponentInstaller and other classes in order to filter, reduce, and apply logic
to items in a collection.

The changes required a change in test expectations for ComponentInstallerTest;
prompts now use the module names being installed in configuration, not the
package name itself.

Additionally, in testing, I discovered that the `AbstractInjector` wasn't
quoting delimiters when preparing the `$isRegisteredPattern` for use with
`preg_match`.