Allow injecting ConfigAggregator-based config files
We recently added zend-config-aggregator, which will supercede
zend-config-manager (developed by a third-party contributor as a
stepping stone). This patch adds support for ConfigAggregator-based
configuration files, while maintaining support for ConfigManager-based
files.

In order to make this work, we needed separate discovery and injector
files. Since both identify the same filesystem location, this also means
an injector chain - and those are intended to run _each_ injector.
Because of that, the existing ExpressiveConfigInjector and the new
ConfigAggregatorInjector now use a new trait, ConditionalDiscoveryTrait,
which makes use of discovery classes in order to determine if
injection/removal is necessary or possible before performing operations.