Default to first discovered option
The primary use case for this plugin is to automate injection of
discovered configuration providers and/or modules. By defaulting to "do
not inject", we are essentially undoing such automation.

This patch modifies the `ComponentInstaller` class to default to the
first provider/module discovered. The implication is that
component/module authors should put the most desired location first if
defining multiple injection points.